Move URLEndpointListener tests to it's own class
