+++ src/Couchbase.Lite.Tests.Shared/P2PTest.cs
@@ -59,7 +59,6 @@ namespace Test
 #endif
     public sealed class P2PTest : TestCase
     {
-        const ushort Port = 0;
         const string ServerCertLabel = "CBL-Server-Cert";
         const string ClientCertLabel = "CBL-Client-Cert";
 
@@ -67,8 +66,6 @@ namespace Test
         private Database _otherDB;
         private Replicator _repl;
         private WaitAssert _waitAssert;
-        URLEndpointListener _listener;
-        URLEndpointListenerConfiguration _config;
 #if !WINDOWS_UWP
         public P2PTest(ITestOutputHelper output) : base(output)
 #else
@@ -373,201 +370,6 @@ namespace Test
         
         #endregion
 
-        #region URLEndpointListener tests
-
-        [Fact]
-        public void TestPort()
-        {
-            int exCnt = 0;
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = Port;
-            _config.DisableTLS = true;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Port.Should().NotBe(0);
-                //stop the listener
-                _listener.Stop();
-                _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
-            }
-        }
-
-        [Fact]
-        public void TestEmptyPort()
-        {
-            int exCnt = 0;
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = 0;
-            _config.DisableTLS = true;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Port.Should().NotBe(0);
-                //stop the listener
-                _listener.Stop();
-                _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
-            }
-        }
-
-        [Fact]
-        public void TestBusyPort()
-        {
-            CouchbasePosixException expectedException = null;
-            var listener = ListenerWithTLS(false, null);
-
-            var config = new URLEndpointListenerConfiguration(_otherDB);
-            config.Port = listener.Port;
-            config.DisableTLS = true;
-            var listener1 = new URLEndpointListener(config);
-
-            try {
-                listener1.Start();
-            } catch (CouchbasePosixException ex) {
-                expectedException = ex;
-            } finally {
-                listener.Stop();
-                listener1.Stop();
-            }
-
-            expectedException.Domain.Should().Be(CouchbaseLiteErrorType.POSIX);
-            expectedException.Error.Should().Be(PosixBase.GetCode(nameof(PosixWindows.EADDRINUSE)));
-        }
-
-        [Fact]
-        public void TestUrls()
-        {
-            int exCnt = 0;
-            var config = new URLEndpointListenerConfiguration(_otherDB);
-            var listener = new URLEndpointListener(config);
-            listener.Urls.Count.Should().Be(0);
-
-            try {
-                listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                listener.Urls.Count.Should().NotBe(0);
-                listener.Stop();
-                listener.Urls.Count.Should().Be(0);
-            }
-        }
-
-        [Fact]
-        public void TestStatus()
-        {
-            int exCnt = 0;
-            HashSet<ulong> maxConnectionCount = new HashSet<ulong>(),
-                maxActiveCount = new HashSet<ulong>();
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = Port;
-            _config.DisableTLS = true;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the listener has not yet started.");
-            _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because no client connection has been established.");
-                _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because no client connection has been established.");
-
-                using (var doc1 = new MutableDocument("doc1"))
-                using (var doc2 = new MutableDocument("doc2")) {
-                    doc1.SetString("name", "Sam");
-                    Db.Save(doc1);
-                    doc2.SetString("name", "Mary");
-                    _otherDB.Save(doc2);
-                }
-
-                var targetEndpoint = new URLEndpoint(new Uri($"{_listener.Urls[0]}".Replace("http", "ws")));
-                var config = new ReplicatorConfiguration(Db, targetEndpoint);
-                using (var repl = new Replicator(config)) {
-                    var waitAssert = new WaitAssert();
-                    var token = repl.AddChangeListener((sender, args) =>
-                    {
-                        waitAssert.RunConditionalAssert(() =>
-                        {
-                            maxConnectionCount.Add(_listener.Status.ConnectionCount);
-                            maxActiveCount.Add(_listener.Status.ActiveConnectionCount);
-
-                            return args.Status.Activity == ReplicatorActivityLevel.Stopped;
-                        });
-                    });
-
-                    repl.Start();
-                    try {
-                        waitAssert.WaitForResult(TimeSpan.FromSeconds(100));
-                    } finally {
-                        repl.RemoveChangeListener(token);
-                    }
-                }
-
-                maxConnectionCount.Max().Should().Be(1);
-                maxActiveCount.Max().Should().Be(1);
-
-                //stop the listener
-                _listener.Stop();
-                _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the connection is stopped.");
-                _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the connection is stopped.");
-            }
-        }
-
-        [Fact]
-        public void TestPasswordAuthenticator()
-        {
-            var auth = new ListenerPasswordAuthenticator((sender, username, password) =>
-            {
-                return username == "daniel" && Encoding.Unicode.GetString(password) == "123";
-            });
-
-            _listener = ListenerWithTLS(false, auth);
-
-            // Replicator - No authenticator
-            var targetEndpoint = new URLEndpoint(new Uri($"{_listener.Urls[0]}".Replace("http", "ws")));
-            var config = new ReplicatorConfiguration(Db, targetEndpoint);
-            RunReplication(config, (int) CouchbaseLiteError.HTTPAuthRequired, CouchbaseLiteErrorType.CouchbaseLite);
-
-            // Replicator - Wrong Credentials
-            config.Authenticator = new BasicAuthenticator("daniel", Encoding.Unicode.GetBytes("456"));
-            RunReplication(config, (int) CouchbaseLiteError.HTTPAuthRequired, CouchbaseLiteErrorType.CouchbaseLite);
-
-            // Replicator - Success
-            config.Authenticator = new BasicAuthenticator("daniel", Encoding.Unicode.GetBytes("123"));
-            RunReplication(config, 0, 0);
-
-            _listener.Stop();
-        }
-
-        #endregion
-
         #region TLSIdentity tests
 
         [Fact]
@@ -582,7 +384,7 @@ namespace Test
         [Fact]
         public void TestCreateDuplicateClientIdentity() => CreateDuplicateServerIdentity(false);
 
-        [Fact]
+        //[Fact]
         public void TestGetIdentityWithCertCollection()
         {
             TLSIdentity id;
@@ -604,7 +406,7 @@ namespace Test
             TLSIdentity.DeleteIdentity(ClientCertLabel).Should().BeTrue();
         }
         
-        [Fact]
+        //[Fact]
         public void TestImportIdentityFromCertCollection()
         {
             TLSIdentity id;
@@ -799,37 +601,6 @@ namespace Test
             return cert.GetPublicKey();
         }
         #endregion
-
-        private URLEndpoint LocalEndPoint(bool tls, Uri uri)
-        {
-            var scheme = tls ? "wss" : "ws";
-            return new URLEndpoint(new Uri($"{uri}".Replace("ws", "wss")));
-        }
-
-        private URLEndpointListener ListenerWithTLS(bool tls, IListenerAuthenticator auth)
-        {
-            int exCnt = 0;
-            
-            var config = new URLEndpointListenerConfiguration(_otherDB);
-            config.DisableTLS = !tls;
-            config.Authenticator = auth;
-
-            var listener = new URLEndpointListener(config);
-
-            try {
-                //start the listener
-                listener.Start();
-            } catch {
-                exCnt++;
-            }
-
-            if (exCnt == 0)
-                return listener;
-            else
-                return null;
-        }
-
-        
 
         private ReplicatorConfiguration CreateFailureP2PConfiguration(ProtocolType protocolType, MockConnectionLifecycleLocation location, bool recoverable)
         {
