+++ src/Couchbase.Lite.Tests.Shared/URLEndpointListenerTest.cs
@@ -22,9 +22,10 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
-
+using System.Text;
 using Couchbase.Lite;
 using Couchbase.Lite.P2P;
 using Couchbase.Lite.Sync;
@@ -70,6 +71,7 @@ namespace Test
     {
         #region Constants
 
+        private const ushort dynamicPort = 0;
         private const ushort WsPort = 4984;
         private const ushort WssPort = 4985;
         private const string ServerCertLabel = "CBL-Server-Cert";
@@ -94,6 +96,142 @@ namespace Test
         #endregion
 
         #region Public Methods
+
+        [Fact]
+        public void TestPort()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            //init and start a listener
+            _listener = CreateListener(false);
+            _listener.Port.Should().Be(WsPort);
+            //stop the listener
+            _listener.Stop();
+            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
+        }
+
+        [Fact]
+        public void TestEmptyPort()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            //init and start a listener
+            var config = CreateListenerConfig(false, null, null, true);
+            _listener = Listen(config, 0, 0);
+
+            _listener.Port.Should().NotBe(0, "Because the port is dynamically assigned.");
+
+            //stop the listener
+            _listener.Stop();
+            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
+        }
+
+        [Fact]
+        public void TestBusyPort()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            var listener = CreateListener(false);
+            //listener1 uses the same port as listener
+            var config = CreateListenerConfig(false);
+            var listener1 = Listen(config, PosixBase.GetCode(nameof(PosixWindows.EADDRINUSE)), CouchbaseLiteErrorType.POSIX);
+
+            listener.Stop();
+            listener1.Stop();
+        }
+
+        [Fact]
+        public void TestUrls()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            var listener = CreateListener(false);
+
+            listener.Urls.Count.Should().NotBe(0);
+            listener.Stop();
+            listener.Urls.Count.Should().Be(0);
+        }
+
+        [Fact]
+        public void TestStatus()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            HashSet<ulong> maxConnectionCount = new HashSet<ulong>(),
+                maxActiveCount = new HashSet<ulong>();
+
+            //init and start a listener
+            _listener = CreateListener(false);
+
+            //listener is started at this point
+            _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because no client connection has been established.");
+            _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because no client connection has been established.");
+
+            using (var doc1 = new MutableDocument("doc1"))
+            using (var doc2 = new MutableDocument("doc2")) {
+                doc1.SetString("name", "Sam");
+                Db.Save(doc1);
+            }
+
+            var targetEndpoint = new URLEndpoint(new Uri($"{_listener.Urls[0]}".Replace("http", "ws")));
+            var config = new ReplicatorConfiguration(Db, targetEndpoint);
+            using (var repl = new Replicator(config)) {
+                var waitAssert = new WaitAssert();
+                var token = repl.AddChangeListener((sender, args) =>
+                {
+                    waitAssert.RunConditionalAssert(() =>
+                    {
+                        maxConnectionCount.Add(_listener.Status.ConnectionCount);
+                        maxActiveCount.Add(_listener.Status.ActiveConnectionCount);
+
+                        return args.Status.Activity == ReplicatorActivityLevel.Stopped;
+                    });
+                });
+
+                repl.Start();
+                try {
+                    waitAssert.WaitForResult(TimeSpan.FromSeconds(100));
+                } finally {
+                    repl.RemoveChangeListener(token);
+                }
+            }
+
+            maxConnectionCount.Max().Should().Be(1);
+            maxActiveCount.Max().Should().Be(1);
+
+            //stop the listener
+            _listener.Stop();
+            _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the connection is stopped.");
+            _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the connection is stopped.");
+        }
+
+        [Fact]
+        public void TestPasswordAuthenticator()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            var auth = new ListenerPasswordAuthenticator((sender, username, password) =>
+            {
+                return username == "daniel" && Encoding.Unicode.GetString(password) == "123";
+            });
+
+            _listener = CreateListener(false, auth);
+
+            // Replicator - No authenticator
+            var targetEndpoint = new URLEndpoint(new Uri($"{_listener.Urls[0]}".Replace("http", "ws")));
+            var config = new ReplicatorConfiguration(Db, targetEndpoint);
+            RunReplication(config, (int) CouchbaseLiteError.HTTPAuthRequired, CouchbaseLiteErrorType.CouchbaseLite);
+
+            // Replicator - Wrong Credentials
+            config.Authenticator = new BasicAuthenticator("daniel", Encoding.Unicode.GetBytes("456"));
+            RunReplication(config, (int) CouchbaseLiteError.HTTPAuthRequired, CouchbaseLiteErrorType.CouchbaseLite);
+
+            // Replicator - Success
+            config.Authenticator = new BasicAuthenticator("daniel", Encoding.Unicode.GetBytes("123"));
+            RunReplication(config, 0, 0);
+
+            _listener.Stop();
+        }
 
         [Fact]
         public void TestServerCertVerificationModeSelfSigned()
@@ -131,16 +269,68 @@ namespace Test
             );
         }
 
+        [Fact]
+        public void TestTLSIdentity()
+        {
+            if (!HasPersistentKeyStorage) return;
+
+            // TLS is disabled
+            _listener = CreateListener(false);
+            _listener.TlsIdentity.Should().BeNull();
+            _listener.Stop();
+            _listener.TlsIdentity.Should().BeNull();
+
+            // Anonymous Identity
+            _listener = CreateListener(true);
+            _listener.TlsIdentity.Should().NotBeNull();
+            _listener.Stop();
+            _listener.TlsIdentity.Should().BeNull();
+
+            // User Identity
+            TLSIdentity.DeleteIdentity(ServerCertLabel).Should().BeTrue();
+            var id = TLSIdentity.CreateIdentity(false,
+                new Dictionary<string, string>() { { Certificate.CommonNameAttribute, "CBL-Server" } },
+                null,
+                ServerCertLabel);
+            var config = CreateListenerConfig(true, null, id);
+            _listener = new URLEndpointListener(config);
+            _listener.TlsIdentity.Should().BeNull();
+            _listener.Start();
+            _listener.TlsIdentity.Should().NotBeNull();
+            _listener.TlsIdentity.Should().BeEquivalentTo(config.TlsIdentity);
+            _listener.Stop();
+            _listener.TlsIdentity.Should().BeNull();
+        }
+
         #endregion
 
         #region Private Methods
+
+        private URLEndpointListenerConfiguration CreateListenerConfig(bool tls = true, 
+            IListenerAuthenticator auth = null, TLSIdentity id = null, bool useDynamicPort = false)
+        {
+            _listener?.Stop();
+
+            var config = new URLEndpointListenerConfiguration(OtherDb);
+            if (useDynamicPort) {
+                config.Port = 0;
+            } else {
+                config.Port = tls ? WssPort : WsPort;
+            }
+
+            config.DisableTLS = !tls;
+            config.Authenticator = auth;
+            config.TlsIdentity = id;
+
+            return config;
+        }
 
         private URLEndpointListener CreateListener(bool tls = true, IListenerAuthenticator auth = null)
         {
             _listener?.Stop();
 
             var config = new URLEndpointListenerConfiguration(OtherDb);
-            config.Port = tls ? WssPort : WsPort;
+            config.Port = tls ? WssPort : WsPort; 
             config.DisableTLS = !tls;
             config.Authenticator = auth;
 
@@ -153,6 +343,13 @@ namespace Test
             _listener?.Stop();
 
             _listener = new URLEndpointListener(config);
+
+            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener has not yet started.");
+            _listener.Urls.Count.Should().Be(0, "Listener's Urls count should be 0 because the listener has not yet started.");
+            _listener.TlsIdentity.Should().BeNull("Listener's TlsIdentity should be null because the listener has not yet started.");
+            _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the listener has not yet started.");
+            _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the listener has not yet started.");
+
             try {
                 _listener.Start();
             } catch (CouchbaseLiteException e) {
